Fix gofmt test.
Previously, the gofmt test was not properly causing Travis to fail, because it
would call run_and_comment on a printf command, which always succeeded. Instead,
we create a function for the gofmt check that outputs on stdout and returns a
meaningful status code, then call run_and_comment on that.

Also fix up a few places that expected TRAVIS_PULL_REQUEST=false, which doesn't
happen when running locally.