de-concurrify activity-monitor
The activity monitor spawns a goroutine only to emit some stats. That
goroutine caused some confusion and some other PRs around this code got more
complicated trying to manage the thread-safety.

But the activity-monitor doesn't really need that goroutine to emit its stats. So, we
simplify by removing two letters and a space.