Forward-date test certificate for WFE test.
The previous cert was causing a test failure because it had a NotBefore date too
far in the past.

This is a stopgap fix to fix the build. The real fix, coming soon, will be to
use a fake clock in the WFE test so we can set it to match the date in the test cert.