Add a test for rollback stability for CA serial numbers.
One of the design considerations for the Boulder-CA serial
number store is that, even in the event of extraordinary
circumstance, we must ensure that we issue monotonically
increasing serial numbers, without gaps and without duplicates.

This test verifies that, in the event of a rollback event during
issuance, that there is no serial number gap.