Add load generator with wfe and ocsp-responder subcomponents
Initial work on this was ripped out of the 'coyote' branch since it's better suited
to being worked on separately.

This adds cmd/load-generator, a tool for generating load on various publicly facing
boulder services. It currently implements a generator for the WFE and one for the
OCSP-Responder.

At the moment no latency information is collected, both generators should implement
HDR Histograms for logging per-endpoint latency. The WFE generator probably hits too
many endpoints to produce detailed latency information vis-a-vis ca-bench (although
maybe not... I should do some quick math on how much memory it'd end up eating) but
the OCSP-Responder generator could definitely.

The python code that current lives in the 'ca-bench' branch that generates the various
latency charts should be *further* generalized and stuck into, say, 'test/bench/latency-charter.py'
so that it can be used by everything.

The 'ca-bench' branch should prob be generalized in a similar way so that it can be used
to load test services per-RPC call (although, guh, arguments will be annoying there).