set an Accept header on VA HTTP requests
This fixes some mysterious breakages that Let's Encrypt users that also
used mod_security on their domains had.

There's some back and forth about whether the mod_security rule is wise,
but that's captured in a mod_security ticket linked from this PR's
ticket.

This patch is a one-line fix with no probable downside. We're not likely
to want to do many more things to satisfy misunderstandings around HTTP
but this seems fine to help our people out.

Fixes #1019.