Fix flakiness in RA test.
The test checked that the ID of a created message was not equal to 33, but that
could happen on Travis depending on the order and number of test cases, leading
to flakiness. Now instead we check that the ID is equal to the object that was
first created, and we also set the overwrite attempt ID high enough that it will
almost certainly never be reached in practice.