issue without a CommonName if not given one
Also, check if the CN is over 64 bytes.

This does not make sure that the Subject is not-empty. That will require
a cfssl change to add the SerialNumber to Subject. We may want to block
on that. There is the question, though, of how to represent the
Subject's SerialNumber. There are no docs specifying that behavior and
certs I've found with a Subject SerialNumber set have had numbers not
all like the cert's actual SerialNumber. Unfortunately, the specs here
are not helpful. Unlike the actual SerialNumber of a certificate which
is specified as a number, the Subject's SerialNumber is specified as a
printable string. That's it.

I've not been able to find examples of empty Subject's on the web but
perhaps others can.

Fixes #40