Ensure FAKE_DNS is passed to boulder container
The patch adds the explicit -e FAKE_DNS=value to the docker run command that runs the boulder container to ensure that the variable is available in the container.

The patch also removes the code that automatically derived FAKE_DNS. As FAKE_DNS was not passed previously, using that code now can break the cases when run-docker.sh was used under assumption that FAKE_DNS should be 127.0.0.1.