Handle errors in EffectiveTLDPlusOne for ratelimits
The only failure conditions for EffectiveTLDPlusOne are (1) the
underlying PublicSuffix function returning nonsensical data (a suffix
longer than the original domain, or the character preceding the suffix
is not a '.'), and (2) the suffix being equal to the domain.

Since we have bigger problems if (1) ever happens, assume errors are
only returned if (2) has occured, and use the public suffix itself for
the ratelimit.

Fixes #1572