option to use another container network
The patch adds support for NET_CONTAINER environment variable in run-docker.sh. When specified, it instructs all boulder containers to use that container network stack, not the default host networking. This is to support a letsencrypt client running in a docker container that does not use host networking. With the option the boulder containers can join that container network and be available over local address there with no expose on the host.

The patch removes -p port specification in docker run commands and that should not be used with --net host or --net container: . In addition I removed unused code to infer hostip. I forgot to do that in #1569 .