wfe: Honors signing algorithm declared in JWK.
Unless it is HMAC based, which this change explictly disallows
for not relying on asymmetric cryptography, making them unsuitable
for veryfing account ownership.

The remaining behavior is unchanged. If no algorithm is declared in the JWK,
it then proceeds to determine one based on the account key type used,
RS256 is used by default for RSA keys. ECDSA algorithms are determined based
on the curve parameters. PS256, PS384 and PS512 are being shadowed
by RSA256 and not verified unless explicitly declared in the JWK.

The end result for LetsEncrypt clients is that, in order to sign HTTP
requests payloads, using account keys; they will be able to use all algorithms
supported by JOSE specs: https://tools.ietf.org/html/rfc7518,
except those based on HMAC.