allow to redirect syslog to stdout
When syslog network is empty, check `STDOUT_LOG` environment variable and print the log to stdout if it is set bypassing syslog API, as expected with container-based setups. This removed the need to install and run rsyslog daemon inside the boulder docker container.

To stay compatible with the current usage of stdoutLogLevel that intructs to copy log to stdout while annotating it with color terminal escapes the patch allow for STDOUT_LOG to be either `plain` to mean colorless log without timestamps or `terminal` when the format is exactly the current format with stdoutLogLevel=7.

The docker container entry script set the variable depending on if the stdout is tty or not.