Statistics Engine - Initial Version (Feature Parity)
This is a direct port of the `dataviz/plot.py` in current use by Let's Encrypt
for statistics generation. Unlike `plot.py` however, this does not format for
any particular graphing library, it instead produces a raw JSON output that
can itself be processed into graphs.

NOTE: This _does not improve_ the queries. That should certainly be done, as
several of them make poor use of indices or produce relatively worthless data.
This commit is merely to reach feature parity, permit a wider audience to
contribute to stats generation, obtain [glorious feedback benefit from teams
of both devs and ops](https://twitter.com/devops_borat). I will work to
improve the queries after we get the approach nailed down and landed.

The design for this feature is that it should be run periodically against a
read-only replica of the Boulder database, producing a JSON output file which
can be transmitted somehow out of the trusted datacenter environment to a
remote server and processed into graphs.

Processing the output of this tool into pretty graphs will happen elsewhere,
outside the responsibility of Boulder. I intended to write the first pass of
that software; I will get started on that after this lands.

Generally, the changes are:

- Adds a new "statistics" read-only user to the DB
- Sets up the Boulder "next" config to include this command
- Adds a "statistics" command to the Boulder package, which can be executed
  similarly to any other Boulder command.
- Re-implement all of the `dataviz/plot.py` SQL queries in Golang
- Implement tests for those queries