Return relative URLs in `/directory` JSON.
Prior to this commit the `/directory` JSON result was built once in
`Handler()` and returned as-is for all requests. Each endpoint URL was
fully qualified as an absolute URL using the `BaseURL` configuration
parameter. This required a configuration change in order to tweak the
origin being used for subsequent request. Returning purely relative URLs
(e.g. `/acme/new-reg` vs `http://localhost:4000/acme/new-reg`) would
break clients that assume absolute paths.

This commit introduces a new behaviour where the `/directory` JSON is
built per-request using the HTTP Host header in place of the `BaseURL`.
Clients will still receive a fully qualified URL in each directory entry
but we gain the ability to more easily control the host without
requiring config changes.

Since the `request.URL` is not populated (Spare the `Path` attribute) we
can not use `request.URL.Scheme` for the initial `http://` vs `https://`
prefix when constructing the URLs and instead differentiate between the
two cases using the `req.TLS` attribute. This is known to not work in
production where nginx terminates TLS and talks HTTP to the WFE. I will
address this in a subsequent commit.