Support `BaseURL` for backwards compatibility.
After discussing with @jsha we decided that supporting the `BaseURL`
behaviour predating this set of commits will make the rollout easier.

As of this commit the `BaseURL` configuration parameter is restored. If
it is set to a non-empty value then `/directory` URLs will be returned
by prefixing the `BaseURL` to each endpoint.

If the `BaseURL` parameter is not empty then the new behaviour using the
host/protocol learned from the request is used.

To ensure this works I restored the original `TestDirectory` unit test
and moved the new table-based test to `TestDirectoryRelative`.

I also noticed that the `Common.BaseURL` parameter from `cmd/config.go`
was unused after the earlier commit removed the one place it was
invoked. This commit removes that unused parameter.