Handle empty `Host` header case for rel dir.
The code to create absolute URLs for the `/directory` endpoints produced
results like `http:///acme/endpoint` when no `BaseURL` was configured
and an empty `Host` header was sent. This commit assumes a default
`localhost` value for the `Host` header to produce
`http://localhost/acme/endpoint` in this circumstance.

Unclear if there is a good way to get the port # for the WFE to include
in addition. E.g. `localhost:4000` for the docker instance vs
`localhost:4300` for tests.