Handle `/directory` marshal error with `sendError`.
This commit changes `relativeDirectory` to return any JSON marshalling
errors. The calling code in the `/directory` handler will return this
marshal error by way of a `ServerInternal` problem using the `sendError`
method.

In theory we should never encounter such a marshalling error since we
are just marshalling simple string types. That said it never hurts to
assume the worst and pipe the exception in. :-)