Restore `BaseURL` common config param.
The `BaseURL` attribute of the `Common` config struct was removed in
earlier commits from this PR. Unfortunately that is premature as this is
the means by which `wfe.BaseURL` is being set currently in prod. Its use
predates a config split.

This commit restores the struct member in `config.go` and restores the
point in `boulder-wfe/main.go` where `wfe.BaseURL` is set to the value
from `c.Common.BaseURL`.