Clean up exported paths.
As @jsha pointed out in review for #1847 there are no need to have the
`NewReg`, `NewAuthz` etc `WebFrontEndImpl` members anymore. Similarly,
the package-level constants `NewRegPath`, etc can be changed to be
unexported. This cleans up the overall WFE package slightly.

This commit removes the unneeded struct members, moves the
`DirectoryEndpoints` map to a local variable in the `/directory` Handler
and unexports the various `Path` constants.