Modify `parseHTTPConnError` to use convenience functions.
In two places the VA was constructing an explicit `ProblemDetails`
object and populating its `Type` field using the return of the
`parseHTTPConnError` function.

This commit changes the signature of the `parseHTTPConnError` function
to accept a details string and return a fully formed `ProblemDetails`
object instead of a `ProblemType`. The return values are constructed
using the convenience functions from `probs.go` instead of direct
instantiation.