Fix `getSerialsIssuedSince` termination criteria.
As documented in Issue #1872 the `getSerialsIssuedSince` `for { }` loop
will never terminate under sufficient issuance load. The behaviour prior
to this commit would only terminate the loop when `sql.ErrNoRows` or
exactly 0 rows were returned by the query. If we sustain sufficient
issuance requests that there are always a few certificates issued since
the last query the loop will never terminate.

This commit changes the loop behaviour such that rather than terminating
when exactly *0* rows are returned, we terminate when a number of rows
less than the configured `batchSize` is returned.