Adds test for #1872.
This commit modifies ocsp-updater/main.go to define a `dbMockAdapter`
interface that defines the gorp DbMap functions used by the OCSPUpdater
functions. This allows mock implementations to be provided in place of
a real DbMap by unit tests.

Using a `dbMockAdapter` that forever returns 1 serial when queried we
can test for the condition that manifests as Issue #1872 to ensure the
fix in 87881e42822976c4b71da064430c98c1fd691b9b is sufficient to ensure
that even when there is always a new serial returned from the DB that
the loop will terminate.