Protect `rlPolicies` access with RWMutex.
Since we're now using a `reloader` to populate `ra.rlPolicies` we need
to make sure that the readers/updaters use a RWMutex to appropriately
guard concurrent updates/reads.

With this commit a new `rlMu` mutex is added. The reloader invoked
`setRateLimitPolicies` acquires a write lock on the `rlMu` mutex. All
other accessors acquire a read lock on the `rlMu` mutex.