Move mutex into RateLimitConfig, add abstraction.
As pointed out in the reviews for #1894 it is awkward to maintain
a RWMutex next to the RateLimitConfig struct. Instead the
RateLimitConfig should be an outer layer of abstraction that contains
a RWMutex and a pointer to a "real" rateLimitConfig that can be switched
out when required.

This commit refactors rate-limits.go to support this model, with
`RateLimitConfig` now becoming an outer wrapper around a pointer to an
unexported `rateLimitConfig`. The individual fields of the old
RateLimitConfig are replaced by accessor functions that use the
`rateLimitConfig` pointer.

Some unit tests outside of the `ratelimit` package were previously
constructing `RateLimitPolicy` objects directly for a mock
`RateLimitConfig`. Since they can not construct the unexported
`rateLimitConfig` I created a `New` method on the exported
`RateLimitConfig` that accepts a `RateLimitPolicy` for each ofn the
`rateLimitConfig` members. Its a little bit awkward and there may be
room for improvement here.

Similarly, since the "empty" value of a `RateLimitConfig` has a nil
`rateLimitConfig` pointer, the RA `NewRegistrationAuthority` calls
`ra.rlPolicies.New()` with 5 blank `RateLimitPolicy` structs to prevent
nil panics when unit tests are using the policy without setting a policy
file.