Replace RateLimitConfig `New` constructor with mutators.
The `New()` constructor of the `RateLimitConfig` to allow test methods
to supply in-memory RateLimitPolicy objects vs. parsing a full config
file was awkward. It required passing 5 separate `RateLimitConfig`
objects and, since Go doesn't support named arguments, being exceedingly
careful which position each was in.

This commit adds a small bit of code duplication to `rate-limits.go`
but avoids the messy `New` constructor usage. By adding individual `Set`
methods that obey the correct locking semantics the test are clearer and
everyone is happy (yay).