Support mocking RateLimitConfigs for unit tests.
Prior to this commit the `ratelimit.RateLimitConfig` struct supported
unit tests specifying `RateLimitPolicy`'s through a series of `Set`
mutator methods. As pointed out in review, it is more "Go-like" to
remove these special methods and instead support mock implementations.

To this end the mutators are removed and `RateLimitConfig` is changed to
an interface definition. A concrete implementation is provided by way of
the `MutexRateLimitConfig`. For the RA unit tests
a `MockRateLimitConfig` is created that satisfies the `RateLimitConfig`
interface and allows direct setting of each `RateLimitPolicy`.