Reuse existing Authentications when possible.
Presently clients may request a new AuthZ be created for a domain that
they have already proved Authentication. This results in unnecessary
bloat in the authentications table.

This commit alters the `NewAuthorization` function of the RA such that
before going through the work of creating a new authz it checks whether
there is a valid authz for that domain/regID already. If there is, then
we short circuit creation and return the existing authz populated with
the previously solved challenges.

Since clients invoking new AuthZ and getting an AuthZ back expect to
then turn around and post updates to the corresponding challenges we
have to relax some checks server side for when a client submits an
update to an already solved challenge. First we consider a challenge
that is already valid to be "sane" in `Challenge.IsSane`. Next, we alter
the ra's `UpdateAuthorization` to return before trying to save the
challenge/authz when it has already been solved.