Add `reuseValidAuthz` config parameter.
Gate the new authz reuse behind a RA config parameter `reuseValidAuthz`.
In the default case (false) the old RA behaviour is maintained (e.g.
requests for a new authz when an existing valid authz exists will create
an additional authz). This commit sets the flag to true in
`boulder-config-next.json` to enable the reuse behaviour for the "next"
config. Constructors are modified in relevant tests to provide a value
for the new config parameter.