Abort UpdateAuthorization when authz is valid.
Previous to this commit UpdateAuthorization was exiting early for
a reused authz when the challenge being submitted was already valid.
This is not quite right and broke when a client was submitting
a different challenge type than was used for the original reused authz
(e.g. getting an authz with http-01 and then later updating the reused
authz for sni-01).

This commit modifies the reused authz behaviour in UpdateAuthorizations
to return early when the overall authz is already valid, regardless of
the specific challenge being sent for update.

Moving this check earlier in the update flow and ignoring the challenge
also allows the earlier change to loosen the challenge `IsSane` function
to be removed. It is no longer neccessary to consider already valid
challenges as "sane" in the `IsSaneForClientOffer` sense.