Adds test for both `reuseAuthZ` states.
This commit adds tests for both `reuseAuthZ = true` and `reuseAuthZ
= false`.

In the case where reuse is enabled the test ensures that after creating
1 valid, finalized authz that a subsequent request to create a new authz
for the same regID/domain will reuse the first authz. The test also
ensures that sending an update to either a valid or pending challenge
on the overall valid reused authz will not produce errors.

In the case where reuse is disabled the test ensures the old behaviour
is maintained. E.g. that after creating a valid finalized authz that
a subsequent request to create a new authz for the same regID/domain
will produce a brand new authz.