Do not reuse near-expiring (<1d) authz.
If an existing valid authorization will expire within 24 hours the RA
should not reuse it when a client asks to create a new authz for the
same reg ID/domain.

This commit modifies the `reuseValidAuthz`-gated behaviour in
`NewAuthorization` to not reuse any authz's that expire within a day.
A unit test is added to explicitly check that a authz expiring in 12
hours will not be reused.