Allow removing registration contact.
The RA `UpdateRegistration` function merges a base registration object
with an update by calling `Registration.MergeUpdate`. Prior to this
commit `MergeUpdate` only allows the updated registration object to
overwrite the `Contact` field of the existing registration if the
updated reg. defines at least one `AcmeURL`. This prevents clients from
being able to outright *remove* the contact associated with an existing
registration.

This commit removes the `len()` check on the `input.Contact` in
`MergeUpdate` to allow the `r.Contact` field to be ovewritten by
a `[]*core.AcmeURL(nil)`. Subsequently clients can now send an empty contacts list in order to remove their reg contact.