Change `Registration.Contact` type to `*[]*core.AcmeURL`.
Since we allow partial updates to registration objects we must be able
to determine whether a reg update was received with *no* `Contacts` or
whether it was received with *empty* `Contacts`. The best way to support
this is by changing the field to be a pointer type and differentiating
between `nil` and `[]*core.AcmeURL{nil}`.

This commit changes the field type, and also updates the numerous
(mostly test related) places where a `[]*core.AcmeURL` was being
constructed to use a `*[]*core.AcmeURL` instead. For places that may now
be dealing with a reg object that may have a nil contacts field we this
commit adds some safety (e.g. returning early if asked to validate a nil
contacts array, using an empty array when asked to convert a nil
contacts field reg into a model for persistence).