Test removing reg. contact more thoroughly.
This commit moves the contact removal test case from
`TestRegistrationUpdate` into its own `TestRegistrationContactUpdate`
function.

The test cases are changed to avoid instantiating objects directly, and
instead to marshal JSON as this is closer to the "real world" conditions
being tested.

An additional test case is added to ensure that when a update without
a contact is sent that no change to the contact field occurs.