Fixes #1446 - invalid message-id generation.
As documented in #1446, the mailer message-id may generate invalid
message-ids.

While in prod configuring the "from" address to a bare email address
means the message-id's being generated are correct it is not an ideal
solution. The code should always generate the correct message ID even
when given a non-bare from address.

As suggested by @rolandshoemaker this commit fixes the underlying issue
by converting the Mailer `from` field from a string to a proper
`mail.Address`. The relevant callers were updated to provide the correct
type and the test was adjusted to match the expected output.